% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{is.webglobe}
\alias{is.webglobe}
\title{Is it a webglobe?}
\usage{
is.webglobe(x)
}
\arguments{
\item{x}{The object that might be a webglobe}
}
\value{
TRUE or FALSE
}
\description{
Determine if an object is a webglobe
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
is.webglobe(wg)
}

}
