% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{+.webglobe}
\alias{+.webglobe}
\title{Send command}
\usage{
\method{+}{webglobe}(wg, x)
}
\arguments{
\item{wg}{Webglobe}

\item{x}{Command to send}
}
\value{
The same webglobe
}
\description{
Send a command to a webglobe
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg + wgclear()
}

}
