% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{print}
\alias{print}
\title{Display a webglobe}
\usage{
print(wg, ...)
}
\arguments{
\item{wg}{The webglobe}

\item{...}{Other arguments}
}
\value{
NA
}
\description{
Delegates printing to appropriate method
}

