% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{print.webglobe}
\alias{print.webglobe}
\title{Display a webglobe}
\usage{
\method{print}{webglobe}(x, ...)
}
\arguments{
\item{x}{The webglobe}

\item{...}{Ignored}
}
\value{
NA
}
\description{
Displays a webglobe. If the webglobe is immediate, then a browser
            window containing it should already be open; in this case, the
            webglobe's address is returned. If the webglobe is not immediate
            then a new browser is open and the cached pipeline is sent to it.
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe()
wg
}

}
