% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgcamcenter}
\alias{wgcamcenter}
\title{Camera: Center on a point}
\usage{
wgcamcenter(lat, lon, alt = NA)
}
\arguments{
\item{lat}{Latitude of the center point, in degrees}

\item{lon}{Longitude of the center point, in degrees}

\item{alt}{Altitude of the center point, in kilometres}
}
\value{
A webglobe command
}
\description{
Centers the camera on a point
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg+wgcamcenter(45,-93,5000)
}

}
