% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgcamhome}
\alias{wgcamhome}
\title{Camera: Send home}
\usage{
wgcamhome()
}
\value{
A webglobe command
}
\description{
Send the camera to its home location
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg+wgcamhome()
}

}
