% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgclear}
\alias{wgclear}
\title{Clear the scene}
\usage{
wgclear()
}
\value{
A webglobe command
}
\description{
Clears everything from the map
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg+wgclear()
}

}
