% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgimmediate}
\alias{wgimmediate}
\title{Immediate mode: On}
\usage{
wgimmediate()
}
\value{
A webglobe command
}
\description{
Turns on immediate mode
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=FALSE)
wg + wgimmediate() #wg is now immediate
}

}
