% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgimmediate_set}
\alias{wgimmediate_set}
\title{Immediate mode: Set}
\usage{
wgimmediate_set(mode)
}
\arguments{
\item{mode}{TRUE or FALSE: TRUE immplies immediate mode on, FALSE implies off}
}
\value{
A webglobe command
}
\description{
Set immediate mode by value
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg + wgimmediate_set(FALSE) #wg is now unimmediate
}

}
