% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgpolygondf}
\alias{wgpolygondf}
\title{Plot long-frame polygons}
\usage{
wgpolygondf(df, fill = NA, alpha = 1, extrude_height = 0,
  stroke = "yellow", stroke_width = 10)
}
\arguments{
\item{df}{Data frame to plot}

\item{fill}{Fill colour name}

\item{alpha}{Alpha (transparency value)}

\item{extrude_height}{Height of the polygon above the surrounding landscape, in TODO}

\item{stroke}{Outline colour (TODO)}

\item{stroke_width}{Outline width (TODO)}
}
\value{
A webglobe command
}
\description{
Plot polygons defined by long-style data frame
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg+wgpolygondf(ggplot2::map_data("usa"),fill="blue",extrude_height=1000)
}

}
