% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgport}
\alias{wgport}
\title{Get webglobe's port}
\usage{
wgport(wg)
}
\arguments{
\item{wg}{Webglobe whose port should be determined}
}
\value{
A number representing the webglobe's port
}
\description{
Determine which port a webglobe is running on
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wgport(webglobe)
}

}
