% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgtitle}
\alias{wgtitle}
\title{Title of webglobe browser window}
\usage{
wgtitle(title)
}
\arguments{
\item{title}{The title to use}
}
\value{
A webglobe command
}
\description{
Changes the tab/window title of the webglobe's browser view
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg+wgtitle("I am the new title!")
}

}
