% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webglobe.R
\name{wgunimmediate}
\alias{wgunimmediate}
\title{Immediate mode: Off}
\usage{
wgunimmediate()
}
\value{
A webglobe command
}
\description{
Turns off immediate mode
}
\examples{
\dontrun{
library(webglobe)
wg<-webglobe(immediate=TRUE)
wg + wgunimmediate() #wg is now unimmediate
}

}
